segundo cambio a valor entero

+++ github.py
@@ -1,5 +1,5 @@
-num1=float(input("Ingrese numero1:"))
-num2=float(input("ingrese numero2"))
+num1=int(input("Ingrese numero1:"))
+num2=int(input("ingrese numero2"))
 
 sumar=num1+num2
